requirements.txt
langchain_core==0.3.72
langchain_openai==0.3.28
langgraph==0.6.2
pydantic==2.11.7
python-dotenv==1.1.1
tavily_python==0.7.10
